Add tests for WPAD, expand other proxy testing
Attempt to determine whether this network provides proxy auto-detection
(WPAD/autoproxy/PAC), and if so, diagnose whether we get good
information and make use of it.
Improve reporting of suggested actions to take.
Shorten curl connect timeouts so we fail faster. Assuming 10 seconds
without completing a TCP connection is long enough to conclude that
something's wrong, and move on to diagnostics more quickly.