Add DNS test
Add a basic test that attempts to send a DNS request to the configured
servers. This should catch e.g. firewalled hosts being configured to use
public nameservers that they can't even reach.

Cleaned up reporting of the commands that were run. This is intended to
make it easier for a system or network administrator to understand the
specific tests being run, and the relevance of their results.